fix(LandingSearch): refactor geocoder code
